// src/types.ts
export type ShortManifestResource =
  | "catalog"
  | "meta"
  | "stream"
  | "subtitles"
  | "addon_catalog";
export type Extra = "search" | "genre" | "skip";
export type ContentType = "movie" | "series" | "channel" | "tv";

export type DefaultConfig = Record<string, any> | undefined;
export type DefaultHandlerExtra = Record<string, any>;

/**
 * Extra properties for catalog handlers
 */
export type CatalogHandlerExtra = {
  /**
   * String to search for in the catalog
   */
  search?: string;

  /**
   * A string to filter the feed or search results by genres
   */
  genre?: string;

  /**
   * Used for catalog pagination, refers to the number of items skipped from the beginning of the catalog.
   * The standard page size in Stremio is 100, so the `skip` value will be a multiple of 100.
   * If you return less than 100 items, Stremio will consider this to be the end of the catalog.
   */
  skip?: number;
};
// ...
export type GetHandlerExtra<T extends ShortManifestResource> =
  T extends keyof HandlerExtraMap ? HandlerExtraMap[T] : DefaultHandlerExtra;

/**
 * Generic handler arguments with conditional `Extra` typing
 */
export interface HandlerArgs<
  TResource extends ShortManifestResource = ShortManifestResource,
  TConfig = DefaultConfig,
  TExtra = GetHandlerExtra<TResource>,
> {
  type: ContentType;
  id: string;
  extra: TExtra;
  config: TConfig;
}

// Specific type aliases for each handler type
export type CatalogHandlerArgs<Config = DefaultConfig> = HandlerArgs<
  "catalog",
  Config,
  CatalogHandlerExtra
>;
export type MetaHandlerArgs<Config = DefaultConfig> = HandlerArgs<
  "meta",
  Config,
  DefaultHandlerExtra
>;
export type StreamHandlerArgs<Config = DefaultConfig> = HandlerArgs<
  "stream",
  Config,
  DefaultHandlerExtra
>;
export type SubtitlesHandlerArgs<Config = DefaultConfig> = HandlerArgs<
  "subtitles",
  Config,
  SubtitlesHandlerExtra
>;
// ...
export type WithCache<T> = T & Cache;

/**
 * Summarized collection of meta items.
 *
 * Catalogs are displayed on the Stremio's Board, Discover and Search.
 */
export interface MetaPreview {
  /**
   * Universal identifier.
   * You may use a prefix unique to your addon.
   *
   * Example: 'yt_id:UCrDkAvwZum-UTjHmzDI2iIw'
   */
  id: string;
  /**
   * Type of the content.
   */
  type: ContentType;
  /**
   * Name of the content.
   */
  name: string;
  /**
   * URL to PNG of poster.
   *
   * Accepted aspect ratios: 1:0.675 (IMDb poster type) or 1:1 (square).
   *
   * You can use any resolution, as long as the file size is below 100kb.
   * Below 50kb is recommended.
   *
   * Note: According to the Meta Preview Object documentation, this should be required for catalog responses,
   * but kept optional here for compatibility with Meta Object documentation.
   */
  poster?: string;
  /**
   * Poster can be square (1:1 aspect) or poster (1:0.675) or landscape (1:1.77).
   *
   * Defaults to 'poster'.
   */
  posterShape?: "square" | "poster" | "landscape";
  /**
   * The background shown on the stremio detail page.
   *
   * Heavily encouraged if you want your content to look good.
   *
   * URL to PNG, max file size 500kb.
   */
  background?: string;
  /**
   * The logo shown on the stremio detail page.
   *
   * Encouraged if you want your content to look good.
   *
   * URL to PNG.
   */
  logo?: string;
  /**
   * A few sentences describing your content.
   */
  description?: string;
  /**
   * Array containing objects in the form of { "source": "P6AaSMfXHbA", "type": "Trailer" }.
   *
   * Where source is a YouTube Video ID and type can be either "Trailer" or "Clip".
   * Used for the Discover Page Sidebar.
   *
   * @deprecated This will soon be deprecated in favor of `meta.trailers` being an array of Stream Objects.
   */
  trailers?: Array<{ source: string; type: "Trailer" | "Clip" }>;
}

/**
 * Detailed description of a meta item.
 *
 * This description is displayed when the user selects an item from the catalog.
 */
export interface MetaDetail extends MetaPreview {
  /**
   * genre/categories of the content.
   *
   * e.g. ["Thriller", "Horror"]
   *
   * **WARNING: this will soon be deprecated, use 'links' instead**
   */
  genres?: string[];
  releaseInfo?: string;
  /**
   * Array of directors.
   *
   * Deprecated: use 'links' instead
   *
   * @deprecated
   */
  director?: string[];
  /**
   * Array of members of the cast.
   *
   * use 'links' instead
   *
   * @deprecated
   */
  cast?: string[];
  /**
   * IMDb rating, which should be a number from 0.0 to 10.0.
   */
  imdbRating?: string;
  /**
   * ISO 8601, initial release date.
   *
   * For movies, this is the cinema debut.
   *
   * e.g. "2010-12-06T05:00:00.000Z"
   */
  released?: string;
  /**
   * Array containing objects in the form of { "source": "P6AaSMfXHbA", "type": "Trailer" }.
   *
   * Where source is a YouTube Video ID and type can be either "Trailer" or "Clip".
   *
   * @deprecated This will soon be deprecated in favor of meta.trailers being an array of Stream Objects.
   */
  trailers?: Array<{ source: string; type: "Trailer" | "Clip" }>;
  /**
   * Can be used to link to internal pages of Stremio.
   *
   * example: array of actor / genre / director links.
   */
  links?: MetaLink[];
  /**
   * Used for channel and series.
   *
   * If you do not provide this (e.g. for movie), Stremio assumes this meta item has one video, and it's ID is equal to the meta item id.
   */
  videos?: MetaVideo[];
  /**
   * Human-readable expected runtime.
   *
   * e.g. "120m"
   */
  runtime?: string;
  /**
   * Spoken language.
   */
  language?: string;
  /**
   * Official country of origin.
   */
  country?: string;
  /**
   * Human-readable that describes all the significant awards.
   */
  awards?: string;
  /**
   * URL to official website.
   */
  website?: string;
  behaviorHints?:
    | {
        /**
         * Set to a Video Object id in order to open the Detail page directly to that video's streams.
         */
        defaultVideoId?: string;
      }
    | undefined;
}
// ...
export interface Subtitle {
  /**
   * Unique identifier for each subtitle, if you have more than one subtitle with the same language, the id will differentiate them.
   */
  id: string;
  /**
   * Url to the subtitle file.
   */
  url: string;
  /**
   * Language code for the subtitle, if a valid ISO 639-2 code is not sent, the text of this value will be used instead.
   */
  lang: string;
  /**
   * Label for the subtitle, used for display purposes.
   */
  label?: string;
}

/**
 * The addon description and capabilities.
 *
 * The first thing to define for your addon is the manifest, which describes it's name, purpose and some technical details.
 */
export interface Manifest {
  /**
   * Identifier, dot-separated, e.g. "com.stremio.filmon"
   */
  id: string;
  /**
   * Human readable name
   */
  name: string;
  /**
   *  Human readable description
   */
  description: string;
  /**
   * Semantic version of the addon
   */
  version: string;
  /**
   * Supported resources, defined as an array of objects (long version) or strings (short version).
   *
   * Example #1: [{"name": "stream", "types": ["movie"], "idPrefixes": ["tt"]}]
   *
   * Example #2: ["catalog", "meta", "stream", "subtitles", "addon_catalog"]
   */
  resources: Array<ShortManifestResource | FullManifestResource>;
  /**
   * Supported types.
   */
  types: string[];
  /**
   * Use this if you want your addon to be called only for specific content IDs.
   *
   * For example, if you set this to ["yt_id:", "tt"], your addon will only be called for id values that start with 'yt_id:' or 'tt'.
   */
  idPrefixes?: string[];
  /**
   * A list of the content catalogs your addon provides.
   *
   * Leave this an empty array ([]) if your addon does not provide the catalog resource.
   */
  catalogs: ManifestCatalog[];
  /**
   * Array of Catalog objects, a list of other addon manifests.
   *
   * This can be used for an addon to act just as a catalog of other addons.
   */
  addonCatalogs?: ManifestCatalog[];

  /**
   * A list of settings that users can set for your addon.
   */
  config?: ManifestConfig[];

  /**
   * Background image for the addon.
   *
   * URL to png/jpg, at least 1024x786 resolution.
   */
  background?: string;

  /**
   * @deprecated use `logo` instead.
   */
  icon?: string;

  /**
   * Logo icon, URL to png, monochrome, 256x256.
   */
  logo?: string;
  /**
   * Contact email for addon issues.
   * Used for the Report button in the app.
   * Also, the Stremio team may reach you on this email for anything relating your addon.
   */
  contactEmail?: string;
  behaviorHints?:
    | {
        /**
         * If the addon includes adult content.
         *
         * Defaults to false.
         */
        adult?: boolean;
        /**
         * If the addon includes P2P content, such as BitTorrent, which may reveal the user's IP to other streaming parties.
         *
         * Used to provide an adequate warning to the user.
         */
        p2p?: boolean;

        /**
         * Default is `false`. If the addon supports settings, it will add a button next to "Install" in Stremio that will point to the `/configure` path on the addon's domain. For more information, read [User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/api/responses/manifest.md#user-data) (or if you are not using the Addon SDK, read: [Advanced User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#using-user-data-in-addons) and [Creating Addon Configuration Pages](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#creating-addon-configuration-pages))
         */
        configurable?: boolean;

        /**
         * Default is `false`. If set to `true`, the "Install" button will not show for your addon in Stremio. Instead a "Configure" button will show pointing to the `/configure` path on the addon's domain. For more information, read [User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/api/responses/manifest.md#user-data) (or if you are not using the Addon SDK, read: [Advanced User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#using-user-data-in-addons) and [Creating Addon Configuration Pages](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#creating-addon-configuration-pages))
         */
        configurationRequired?: boolean;
      }
    | undefined;

  stremioAddonsConfig?: StremioAddonsConfig;
}

export type ManifestConfigType =
  | "text"
  | "number"
  | "password"
  | "checkbox"
  | "select";

/**
 * Addon setting.
 */
export interface ManifestConfig {
  /**
   * A key that will identify the user chosen value.
   */
  key: string;

  /**
   * The type of data that the setting stores.
   */
  type: ManifestConfigType;

  /**
   * The default value. For `type: "boolean"` this can be set to "checked" to default to enabled.
   */
  default?: string;

  /**
   * The title of the setting.
   */
  title?: string;

  /**
   * List of (string) choices for `type: "select"`
   */
  options?: string[];

  /**
   * If the value is required or not. Only applies to the following types: "string", "number". (default is `false`)
   */
  required?: boolean;
}

/**
 * Used as a response for defineResourceHandler.
 */
export interface AddonCatalog {
  /**
   * only http is currently officially supported.
   */
  transportName: string;
  /**
   * The URL of the addon's manifest.json file.
   */
  transportUrl: string;
  /**
   * Object representing the addon's Manifest Object.
   */
  manifest: Manifest;
}

export interface FullManifestResource {
  /**
   * Resource name.
   */
  name: ShortManifestResource;
  /**
   * Supported types.
   */
  types: string[];
  /**
   * Use this if you want your addon to be called only for specific content IDs
   *
   * For example, if you set this to ["yt_id:", "tt"], your addon will only be called for id values that start with 'yt_id:' or 'tt'.
   */
  idPrefixes?: string[];
}

export interface ManifestCatalog {
  /**
   *  This is the content type of the catalog.
   */
  type: string;
  /**
   * The id of the catalog, can be any unique string describing the catalog (unique per addon, as an addon can have many catalogs).
   *
   * For example: if the catalog name is "Favourite Youtube Videos", the id can be "fav_youtube_videos".
   */
  id: string;
  /**
   * Human readable name of the catalog.
   */
  name: string;
  /**
   * Use the 'options' property of 'extra' instead.
   * @deprecated
   */
  genres?: string[];
  /**
   * All extra properties related to this catalog.
   */
  extra?: ManifestExtra[];
}
// ...
export interface StremioAddonsConfig {
  issuer: "https://stremio-addons.net";
  signature: string;
}

/**
 * The addonInterface, as returned from builder.getInterface()
 */
export interface AddonInterface {
  manifest: Manifest;
  get: (
    resource: ShortManifestResource,
    type: ContentType,
    id: string,
    extra?: Record<string, any>,
    config?: Record<string, any>,
  ) => Promise<any>;
}

export type ManifestSchema = StandardSchemaV1<Manifest>;